pyproject.toml
[project]
name = "agent-042"
version = "0.1.0"
description = "AI agent for AI researchers"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "numpy>=2.3.5",
    "omegaconf>=2.3.0",
    "pre-commit>=4.5.0",
    "torch>=2.9.1",
    "transformers>=4.44.0",
    "datasets>=3.0.1",
    "peft>=0.13.0",
    "pytorch-lightning>=2.4.0",
    "hf-xet>=1.2.0",
    "bitsandbytes>=0.48.2",
    "ruff>=0.14.8",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["I", "E", "F"]
